Update the main loop to redraw screen on timer
I've set the timer to run 10 times a second.
We always check the buttons every loop, but
we don't need to redraw the screen unless the state
has changed in some form, so don't bother.

This helps prevent the flickering, which we should sort later.
The right way to do this is to either double buffer or use the
TEAR pin.  With the enable_tear call, when the screen starts
drawing, the TEAR pin will go low, and we should initiate a draw.
That would sync memory copy to the draw method.