Implement score and crashing into pipes
Score is pretty simple, each time we pass a pipe location, increment
the score by one.
I wanted to display the score on the end screen so made it global though
:(

Crashing into pipes was a lot harder.  Refactored how we track the birds
x coordinate more clearly, and then started working out if the bird is
currently between two pipes, and whether the birds y coordinate is in
the gap between the pipes.

Also added a bit of debugging text, changed how high the bird jumps and
rearranged the pipes so it was easier to debug whether I was checking
the right pipe.