Randomise pipes and high score
Randomising the pipes isn't too hard.  Instead of a list of
pipe-heights, we simply generate random pipe_heights at the beginning of
the game.
Sadly, the random module is not included in micropython it seems, so I
had to implement my own randint function.  This is not truely random and
almost certainly exhibits a bias, but it's a start.

To make the game not impossible, I've made each number be within a bound
of the previous number so you can't get a pipe high at 5 and then at 22,
which would be almost impossible to beat.
An interesting idea might to make that gap a function of the distance
travelled, so the game gets harder.
It's also not infinite yet, it creates 50 pipes and you'll just end if
you get there.

Secondly, I added a high score counter.  This keeps the high score in
the database (config.json) which is perfectly fine for now.

I think I'm ready to publish on the basis of this.